Move duplicated logic from Esi/Ssi to an AbstractSurrogate
Fabbot fixes

Move duplicated logic from Esi/Ssi::process in AbstractSurrogate::removeFromControl

Fabbot fixes

Fix comment
